requirements.txt
psycopg2
google-genai
